added user validation

+++ app/controllers/analog_synthesizers_controller.rb
@@ -8,36 +8,45 @@ class AnalogSynthesizerController < ApplicationController
 
     # new
     get "/analog_synthesizers/new" do
-         erb :"analog_synthesizers/new.html"
+        redirect_if_not_logged_in
+        erb :"analog_synthesizers/new.html"
     end
 
     # show
     get "/analog_synthesizers/:id" do
+        redirect_if_not_logged_in
         @as = AnalogSynthesizer.find(params[:id])
         erb :"analog_synthesizers/show.html"
     end
 
     # create
     post "/analog_synthesizers" do
+        redirect_if_not_logged_in
         as = current_user.analog_synthesizers.create(params)
         redirect "/analog_synthesizers/#{as.id}"
     end
 
     # edit
     get "/analog_synthesizers/:id/edit" do
+        redirect_if_not_logged_in
         @as = AnalogSynthesizer.find(params[:id])
         erb :"analog_synthesizers/edit.html"
     end
 
     patch "/analog_synthesizers/:id" do
         as = AnalogSynthesizer.find(params[:id])
-        as.update(params[:analog_synthesizer])
-        redirect "/analog_synthesizers/#{as.id}"
+        if as.user == current_user
+            as.update(params[:analog_synthesizer])
+            redirect "/analog_synthesizers/#{as.id}"
+        end
     end
 
+    # delete
     delete "/analog_synthesizers/:id" do
         as = AnalogSynthesizer.find(params[:id])
-        as.destroy
-        redirect "/analog_synthesizers"
+        if as.user == current_user
+            as.destroy
+            redirect "/analog_synthesizers"
+        end
     end
 end
